HBP-90. Move REST endpoints for HPC in a separate controller.
